requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
coverage==4.5.4
coveralls==1.9.2
Django==3.0.1
django-cors-headers==3.2.0
django-environ==0.4.5
django-extensions==2.2.5
djangorestframework==3.11.0
docopt==0.6.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wrapt==1.11.2
